Auto Playlist Music
This is a list of all the music in the Auto Playlist. If you would like a song added to it, PM Brocky on Discord with the Artist, Name, and Link of the song.

If you are wanting multiple songs, send Brocky a txt file with the above information for all the songs.